refactor(job-store): Honoベストプラクティスに従ったエンドポイント構造の再設計
- Honoの推奨パターンに従いエンドポイントを論理的にグループ化
  - jobFetch/ + jobInsert → job/ に統合（単一求人操作）
  - jobList/ → jobs/ に統合（求人一覧操作）
  - 継続ページネーション機能を jobs/continue/ として分離
- RESTfulな設計への改善:
  - POST /api/v1/job - 求人登録
  - GET /api/v1/job/:jobNumber - 求人詳細取得
  - GET /api/v1/jobs - 求人一覧取得
  - GET /api/v1/jobs/continue - 継続ページネーション
- テスト容易性の向上:
  - 共通ユーティリティをutil.tsに分離
  - DBクライアント注入パターンの準備（今後のモック対応）
- app.tsのルーティング設定を新構造に対応
- 既存のエラーハンドリングとバリデーション機能を維持

参考: https://hono.dev/docs/guides/best-practices#building-a-larger-application
