+++ packages/job-store/src/endpoint/util.ts
@@ -1,0 +1,5 @@
+import * as v from "valibot";
+
+export const envSchema = v.object({
+  JWT_SECRET: v.string(), // JWTSecret の zod schemaがあればそれを使う
+});
